refactor: update dashboard modals to support multiple collateral currencies
