+++ src/utils/helpers/currencies.ts
